Move odogenerator and unions packages into kclient
