+++ pkg/kclient/mock_Client.go
@@ -14,8 +14,8 @@ import (
 	v1 "github.com/openshift/api/project/v1"
 	v10 "github.com/openshift/api/route/v1"
 	v1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
+	unions "github.com/redhat-developer/odo/pkg/kclient/unions"
 	log "github.com/redhat-developer/odo/pkg/log"
-	unions "github.com/redhat-developer/odo/pkg/unions"
 	v11 "k8s.io/api/apps/v1"
 	v12 "k8s.io/api/core/v1"
 	meta "k8s.io/apimachinery/pkg/api/meta"
